+++ src/CfdiUtils/Elements/Cce20/Propietario.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace CfdiUtils\Elements\Cce20;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class Propietario extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'cce20:Propietario';
+    }
+}